shourya-erp/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
jinja2==3.1.4
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
itsdangerous==2.2.0
starlette==0.37.2
aiofiles==23.2.1
